Export `algebra-core::error`
