+++ algebra-core/src/serialize/error.rs
@@ -17,36 +17,10 @@ pub enum SerializationError {
 }
 
 #[cfg(feature = "std")]
-pub(crate) use std::error::Error;
-
-#[cfg(feature = "std")]
-impl Error for SerializationError {
-    fn source(&self) -> Option<&(dyn Error + 'static)> {
-        None
-    }
-}
+impl crate::Error for SerializationError {}
 
 #[cfg(not(feature = "std"))]
-pub trait Error: fmt::Debug + fmt::Display {}
-
-#[cfg(not(feature = "std"))]
-impl<'a, E: Error + 'a> From<E> for crate::Box<dyn Error + 'a> {
-    fn from(err: E) -> crate::Box<dyn Error + 'a> {
-        crate::Box::new(err)
-    }
-}
-
-#[cfg(not(feature = "std"))]
-impl<T: Error> Error for crate::Box<T> {}
-
-#[cfg(not(feature = "std"))]
-impl Error for SerializationError {}
-
-#[cfg(not(feature = "std"))]
-impl Error for io::Error {}
-
-#[cfg(not(feature = "std"))]
-impl Error for crate::String {}
+impl crate::Error for SerializationError {}
 
 impl From<io::Error> for SerializationError {
     fn from(e: io::Error) -> SerializationError {
